refactor: Перенёс NewCompilationDto, CompilationDtoResponse, UpdateCompilationRequest
