+++ ewm-main-service/src/main/java/ru/practicum/compilation/service/impl/AdminCompilationServiceImpl.java
@@ -8,9 +8,9 @@ import ru.practicum.compilation.mapper.CompilationMapper;
 import ru.practicum.compilation.repository.CompilationRepository;
 import ru.practicum.compilation.service.AdminCompilationService;
 import ru.practicum.compilation.service.CompilationService;
-import ru.practicum.dto.compilation.CompilationDtoResponse;
-import ru.practicum.dto.compilation.NewCompilationDto;
-import ru.practicum.dto.compilation.UpdateCompilationRequest;
+import ru.practicum.compilation.dto.CompilationDtoResponse;
+import ru.practicum.compilation.dto.NewCompilationDto;
+import ru.practicum.compilation.dto.UpdateCompilationRequest;
 import ru.practicum.dto.event.EventShortDto;
 import ru.practicum.event.model.Event;
 import ru.practicum.event.service.EventService;
